# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Calder API",
    "version": "1.0.0",
    "description": "Email infrastructure: send, track, verify. Test keys simulate, never deliver.",
    "contact": { "name": "Calder", "url": "https://calder.click" }
  },
  "servers": [{ "url": "https://api.calder.click" }],
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A Calder secret key: calder_sk_test_<random> or calder_sk_live_<random>. The raw secret is shown only at creation or rotation."
      }
    },
...
  },
  "paths": {
...
    "/v1/senders/{id}/test": {
      "post": {
        "summary": "Test send (reports acceptance only)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "202": { "description": "Queued" } }
      }
    },
...
    "/v1/templates": {
      "get": {
        "summary": "List templates",
        "parameters": [
          { "name": "cursor", "in": "query", "schema": { "type": "string", "maxLength": 200 } },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
          }
        ],
        "responses": { "200": { "description": "Templates and cursor pagination" } }
      },
      "post": {
        "summary": "Create a template (v1 when content given)",
        "responses": { "201": { "description": "Template" } }
      }
    },
    "/v1/templates/{id}": {
      "get": {
        "summary": "Get a template",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Template" } }
      },
      "delete": {
        "summary": "Delete a template",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
    "/v1/templates/{id}/versions": {
      "get": {
        "summary": "List versions",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "cursor", "in": "query", "schema": { "type": "string", "maxLength": 200 } },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
          }
        ],
        "responses": { "200": { "description": "Versions" } }
      },
      "post": {
        "summary": "New auto-numbered version (rollback = post old content)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "201": { "description": "Version" } }
      }
    },
...
  }
}